add func

+++ backend-biz/src/main/java/com/langtuo/teamachine/biz/service/user/RoleMgtServiceImpl.java
@@ -21,6 +21,8 @@ import com.langtuo.teamachine.internal.util.LocaleUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
 
 import javax.annotation.Resource;
@@ -45,6 +47,7 @@ public class RoleMgtServiceImpl implements RoleMgtService {
     private AdminAccessor adminAccessor;
 
     @Override
+    @Transactional(readOnly = true)
     public TeaMachineResult<RoleDTO> getByCode(String tenantCode, String roleCode) {
         RolePO rolePO = roleAccessor.getByRoleCode(tenantCode, roleCode);
         RoleDTO roleDTO = RoleMgtConvertor.convertToRoleDTO(rolePO);
@@ -52,6 +55,7 @@ public class RoleMgtServiceImpl implements RoleMgtService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public TeaMachineResult<PageDTO<RoleDTO>> search(String tenantCode, String roleName, int pageNum, int pageSize) {
         pageNum = pageNum < CommonConsts.MIN_PAGE_NUM ? CommonConsts.MIN_PAGE_NUM : pageNum;
         pageSize = pageSize < CommonConsts.MIN_PAGE_SIZE ? CommonConsts.MIN_PAGE_SIZE : pageSize;
@@ -67,6 +71,7 @@ public class RoleMgtServiceImpl implements RoleMgtService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public TeaMachineResult<List<RoleDTO>> list(String tenantCode) {
         TeaMachineResult<List<RoleDTO>> teaMachineResult;
         try {
@@ -98,84 +103,76 @@ public class RoleMgtServiceImpl implements RoleMgtService {
 
         RolePO rolePO = convertToRolePO(request);
         List<RoleActRelPO> roleActRelPOList = convertRoleActRel(request);
-        if (request.isPutNew()) {
-            return putNew(rolePO, roleActRelPOList);
-        } else {
-            return putUpdate(rolePO, roleActRelPOList);
+        try {
+            if (request.isPutNew()) {
+                return doPutNew(rolePO, roleActRelPOList);
+            } else {
+                return doPutUpdate(rolePO, roleActRelPOList);
+            }
+        } catch (Exception e) {
+            log.error("roleMgtService|put|fatal|" + e.getMessage(), e);
+            return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.DB_ERR_UPDATE_FAIL));
         }
     }
 
-    private TeaMachineResult<Void> putNew(RolePO po, List<RoleActRelPO> actRelPOList) {
-        try {
-            RolePO exist = roleAccessor.getByRoleCode(po.getTenantCode(), po.getRoleCode());
-            if (exist != null) {
-                return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.BIZ_ERR_OBJECT_CODE_DUPLICATED));
-            }
+    @Transactional(rollbackFor = Exception.class, propagation = Propagation.REQUIRED)
+    private TeaMachineResult<Void> doPutNew(RolePO po, List<RoleActRelPO> actRelPOList) {
+        RolePO exist = roleAccessor.getByRoleCode(po.getTenantCode(), po.getRoleCode());
+        if (exist != null) {
+            return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.BIZ_ERR_OBJECT_CODE_DUPLICATED));
+        }
 
-            int inserted = roleAccessor.insert(po);
+        int inserted = roleAccessor.insert(po);
+        if (CommonConsts.DB_INSERTED_ONE_ROW != inserted) {
+            log.error("roleMgtService|putNewRole|error|" + inserted);
+            return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.DB_ERR_INSERT_FAIL));
+        }
+
+        int deleted4RoleActRel = roleActRelAccessor.deleteByRoleCode(po.getTenantCode(), po.getRoleCode());
+        for (RoleActRelPO actRelPO : actRelPOList) {
+            int inserted4actRel = roleActRelAccessor.insert(actRelPO);
             if (CommonConsts.DB_INSERTED_ONE_ROW != inserted) {
-                log.error("roleMgtService|putNewRole|error|" + inserted);
-                return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.DB_ERR_INSERT_FAIL));
+                log.error("roleMgtService|putNewActRel|error|" + inserted4actRel);
             }
-
-            int deleted4RoleActRel = roleActRelAccessor.deleteByRoleCode(po.getTenantCode(), po.getRoleCode());
-            for (RoleActRelPO actRelPO : actRelPOList) {
-                int inserted4actRel = roleActRelAccessor.insert(actRelPO);
-                if (CommonConsts.DB_INSERTED_ONE_ROW != inserted) {
-                    log.error("roleMgtService|putNewActRel|error|" + inserted4actRel);
-                }
-            }
-            return TeaMachineResult.success();
-        } catch (Exception e) {
-            log.error("roleMgtService|putUpdate|fatal|" + e.getMessage(), e);
-            return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.DB_ERR_UPDATE_FAIL));
         }
+        return TeaMachineResult.success();
     }
 
-    private TeaMachineResult<Void> putUpdate(RolePO po, List<RoleActRelPO> actRelPOList) {
-        try {
-            RolePO exist = roleAccessor.getByRoleCode(po.getTenantCode(), po.getRoleCode());
-            if (exist == null) {
-                return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.DB_ERR_UPDATE_FAIL));
-            }
-            if (CommonConsts.ROLE_CODE_TENANT_SUPER.equals(po.getRoleCode())) {
-                AdminPO adminPO = DaoUtils.getAdminPOByLoginSession(po.getTenantCode());
-                if (!CommonConsts.ROLE_CODE_SYS_SUPER.equals(adminPO.getRoleCode())) {
-                    return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(
-                            ErrorCodeEnum.BIZ_ERR_CANNOT_MODIFY_TENANT_SUPER_ADMIN_ROLE));
-                }
-            }
-
-            int updated = roleAccessor.update(po);
-            if (CommonConsts.DB_UPDATED_ONE_ROW != updated) {
-                log.error("roleMgtService|putUpdateRole|error|" + updated);
-                return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.DB_ERR_INSERT_FAIL));
-            }
-
-            int deleted4RoleActRel = roleActRelAccessor.deleteByRoleCode(po.getTenantCode(), po.getRoleCode());
-            for (RoleActRelPO actRelPO : actRelPOList) {
-                int inserted4actRel = roleActRelAccessor.insert(actRelPO);
-                if (CommonConsts.DB_INSERTED_ONE_ROW != inserted4actRel) {
-                    log.error("roleMgtService|putUpdateActRel|error|" + inserted4actRel);
-                }
-            }
-            return TeaMachineResult.success();
-        } catch (Exception e) {
-            log.error("roleMgtService|putUpdate|fatal|" + e.getMessage(), e);
+    @Transactional(rollbackFor = Exception.class, propagation = Propagation.REQUIRED)
+    private TeaMachineResult<Void> doPutUpdate(RolePO po, List<RoleActRelPO> actRelPOList) {
+        RolePO exist = roleAccessor.getByRoleCode(po.getTenantCode(), po.getRoleCode());
+        if (exist == null) {
             return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.DB_ERR_UPDATE_FAIL));
         }
+        if (CommonConsts.ROLE_CODE_TENANT_SUPER.equals(po.getRoleCode())) {
+            AdminPO adminPO = DaoUtils.getAdminPOByLoginSession(po.getTenantCode());
+            if (!CommonConsts.ROLE_CODE_SYS_SUPER.equals(adminPO.getRoleCode())) {
+                return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(
+                        ErrorCodeEnum.BIZ_ERR_CANNOT_MODIFY_TENANT_SUPER_ADMIN_ROLE));
+            }
+        }
+
+        int updated = roleAccessor.update(po);
+        if (CommonConsts.DB_UPDATED_ONE_ROW != updated) {
+            log.error("roleMgtService|putUpdateRole|error|" + updated);
+            return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.DB_ERR_INSERT_FAIL));
+        }
+
+        int deleted4RoleActRel = roleActRelAccessor.deleteByRoleCode(po.getTenantCode(), po.getRoleCode());
+        for (RoleActRelPO actRelPO : actRelPOList) {
+            int inserted4actRel = roleActRelAccessor.insert(actRelPO);
+            if (CommonConsts.DB_INSERTED_ONE_ROW != inserted4actRel) {
+                log.error("roleMgtService|putUpdateActRel|error|" + inserted4actRel);
+            }
+        }
+        return TeaMachineResult.success();
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class, propagation = Propagation.REQUIRED)
     public TeaMachineResult<Void> delete(String tenantCode, String roleCode) {
         if (StringUtils.isBlank(tenantCode) || StringUtils.isBlank(roleCode)) {
             return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.BIZ_ERR_ILLEGAL_ARGUMENT));
-        }
-
-        int adminCount = adminAccessor.countByRoleCode(tenantCode, roleCode);
-        if (adminCount > 0) {
-            return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(
-                    ErrorCodeEnum.BIZ_ERR_CANNOT_DELETE_USING_OBJECT));
         }
 
         if (CommonConsts.ROLE_CODE_TENANT_SUPER.equals(roleCode)) {
@@ -188,17 +185,19 @@ public class RoleMgtServiceImpl implements RoleMgtService {
             }
         }
 
-        TeaMachineResult<Void> teaMachineResult;
         try {
+            int adminCount = adminAccessor.countByRoleCode(tenantCode, roleCode);
+            if (adminCount > 0) {
+                return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(
+                        ErrorCodeEnum.BIZ_ERR_CANNOT_DELETE_USING_OBJECT));
+            }
+
             int deleted = roleAccessor.deleteByRoleCode(tenantCode, roleCode);
             int deleted4Rel = roleActRelAccessor.deleteByRoleCode(tenantCode, roleCode);
-            teaMachineResult = TeaMachineResult.success();
+            return TeaMachineResult.success();
         } catch (Exception e) {
             log.error("roleMgtService|delete|fatal|" + e.getMessage(), e);
-            teaMachineResult = TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.DB_ERR_INSERT_FAIL));
+            return TeaMachineResult.error(LocaleUtils.getErrorMsgDTO(ErrorCodeEnum.DB_ERR_INSERT_FAIL));
         }
-        return teaMachineResult;
     }
-
-
 }
